Improve files structure in api
Moved all functions from routes to controllers. Simplified routes folder
