+++ fact-bounty-server/routes/api/story.js
@@ -1,58 +1,27 @@
-const Story = require('../../models/story')
+var express =  require('express')
 
-// @desc Get all stories
-// @access Public
-exports.storyAll = function (req, res, next) {
-	Story.find({}).then(posts => {
-		return res.status(200).json(posts)
-	})
-}
+const router = express.Router()
 
-// @desc Get stories in a range
-// @access Public
-exports.storyRange = function (req, res, next) {
-	const options = {
-		page: req.params.page,
-		limit: 3
-	}
-	Story.paginate({}, options, function (err, result) {
-		return res.status(200).json(result)
-	})
-}
+const storyController = require('../../controllers/story') 
 
+router.get(
+  '/all', 
+  storyController.storyAll)
 
-// @desc Change upvote count of a story
-// @access Public
-exports.storyChangeUpvoteCount = function (req, res, next) {
-	Story.updateOne({ _id: req.body.story_id }, {
-		$inc: {
-			approved_count: req.body.change_val
-		}
-	}, function (mon_err, affected, mon_res) {
-		return mon_err ? res.status(200).json(mon_err) : res.status(200).send()
-	})
-}
+router.get(
+  '/get-range/:page', 
+  storyController.storyRange)
 
-// @desc Change downvote count of a story
-// @access Public
-exports.storyChangeDownvoteCount = function (req, res, next) {
-	Story.updateOne({ _id: req.body.story_id }, {
-		$inc: {
-			fake_count: req.body.change_val
-		}
-	}, function (mon_err, affected, mon_res) {
-		return mon_err ? res.status(200).json(mon_err) : res.status(200).send()
-	})
-}
+router.post(
+  '/change-upvote-count', 
+  storyController.storyChangeUpvoteCount)
 
-// @desc Change mixedvote count of a story
-// @access Public
-exports.storyChangeMixedvoteCount = function (req, res, next) {
-	Story.updateOne({ _id: req.body.story_id }, {
-		$inc: {
-			mixedvote_count: req.body.change_val
-		}
-	}, function (mon_err, affected, mon_res) {
-		return mon_err ? res.status(200).json(mon_err) : res.status(200).send()
-	})
-}
+router.post(
+  '/change-downvote-count', 
+  storyController.storyChangeDownvoteCount)
+
+router.post(
+  '/change-mixedvote-count', 
+  storyController.storyChangeDownvoteCount)
+
+module.exports = router
